requirements.dev.txt
ipython
ipdb
pudb
pytest-watch
click
rich-click

# Linting - Code quality
flake8
pyproject-flake8
black
isort
